+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Added default template files .codecov.yml, .gitattributes, and .gitignore, and
   .vscode folder.
 - Add Unit Tests for xExchAntiMalwareScanning
+- Add remaining Unit Tests for xExchInstall, and for most common setup
+  functions
 
 ## 1.23.0.0